feat(skip_methods?): Handle relative paths and cache result
I was running ./performance/benchmark.rb (which traces a simple recursive function) from the codebeacon_tracer root dir. In this case the path of code traced in benchmark is the relative string I passed in rather than the abs path. As a result it didn't match the configured root and all traces were excluded. Converting the path to absolute all of the time caused problems excluding "<internal..." style paths which are on the exclusion list, but when converted to absolute were misinterpreted as files in the root_path. These got matched and incorrectly traced. This method needs to be very performant. It will be called many more times that the actual tracer functions, so I finally added that caching I've been dreaming about.