fix(config): fix config paths and add (eval) to exclusion list
config.yml was not 1. not being found and 2. probably not in the optimal location. It has been moved to a better location.
lib_root and config_path are internal to this gem but were not reliably being calculated so I added a robust gem_root_path calculation to find this gem root